ENG-1044/ENG-1045: implementation of fault tolerant paged search.
git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@24597 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ dso-l1/src/main/java/com/tc/object/servermap/localcache/impl/ServerMapLocalCacheImpl.java
@@ -181,7 +181,7 @@ public final class ServerMapLocalCacheImpl implements ServerMapLocalCache {
       if (txn == null) { throw new UnlockedSharedObjectException(
                                                                  "Attempt to access a shared object outside the scope of a shared lock.",
                                                                  Thread.currentThread().getName(), manager
-                                                                     .getClientID()); }
+                                                                     .getClientID().toLong()); }
       keyToListeners.put(key, listener);
       txn.addTransactionCompleteListener(listener);
     }
